Bert APi - evaluation/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
command-not-found==0.3
confection==0.1.5
cryptography==44.0.0
cymem==2.0.11
datasets==3.2.0
dbus-python==1.2.18
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
httplib2==0.20.2
huggingface-hub==0.27.1
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
langcodes==3.5.0
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meson==1.6.1
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
netifaces==0.11.0
networkx==3.2.1
numpy==2.2.2
oauthlib==3.2.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pillow==11.1.0
preshed==3.0.9
propcache==0.2.1
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
pytesseract==0.3.13
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
systemd-python==234
thinc==8.3.4
tokenizers==0.21.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0
